tests/wpt/web-platform-tests/feature-policy/reporting/vr-report-only.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
  </head>
  <body>
    <script>
const check_report_format = ([reports, observer]) => {
  const report = reports[0];
  assert_equals(report.type, "feature-policy");
  assert_equals(report.url, document.location.href);
  assert_equals(report.body.feature, "vr");
  assert_equals(report.body.disposition, "report");
  assert_equals(report.body.sourceFile, document.location.href);
  assert_equals(typeof report.body.message, "string");
  assert_equals(typeof report.body.lineNumber, "number");
  assert_equals(typeof report.body.columnNumber, "number");
};

promise_test(async t => {
  const report = new Promise(resolve => {
    new ReportingObserver((reports, observer) => resolve([reports, observer]),
                          {types: ['feature-policy']}).observe();
  });
  await navigator.getVRDisplays();
  check_report_format(await report);
}, "VR report only mode");
    </script>
  </body>
</html>
